pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incident-commander-env"
version = "0.1.0"
description = "OpenEnv-style multi-agent software incident response environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "matplotlib>=3.8.0",
    "openenv-core[core]>=0.2.3",
]

[project.optional-dependencies]
training = [
    "trl>=0.24.0",
    "unsloth",
    "datasets>=3.0.0",
    "transformers>=4.50.0",
]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "incident_commander_env.server.app:main"
incident-demo = "scripts.run_demo:main"
incident-eval = "scripts.evaluate_baseline:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["incident_commander_env*", "server"]
